requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
pyyaml>=6.0
onnx>=1.12.0
onnxruntime>=1.13.0
tqdm>=4.64.0
